Upload in msg directory
Uploaded files are getting stored in a folder with the corresponding chatRecord id as a name and in that folder in another folder corresponding with the id of the msg.

This is needed to prevent overwriting files with same names in a given directory. Just 1 file to upload per message is allowed