Working up and download of files
On adding a message through the firestoreService it checks for a attached file to upload. If provided, the storage Service upload the file to the cloud storage and saves the file in a directory that can be reference down to the message document id from the Firestore via the chatRecord id. Then it retrieves the url to "download" the file from the storage and stores it directly into the created message class object