Fix leading msg
The preview pictures of the file are now being shown correctly in the thread leading message.

Shifted HTML structure around to scroll the record div and the leading msg div. Separating them is possible to stick the leading msg to the top, but needs further adjustment when there is along text or picture present that would push the scrollable Chatrecord out of view.